Major overhaul to AeroVeloz system:
- Added DTOs, service contracts, and event handlers for flights, subscriptions, and authentication in the application layer.
- Refined domain entities, enums, and models; centralized domain validation and error codes.
- Implemented repositories, domain services, notification channels (Email, SMS, OneSignal), and monitoring in infrastructure.
- Introduced IOC project for DI and modular service registration.
- Updated web layer for JWT auth and service wiring.
- Cleaned up obsolete code, improved documentation, and ensured modern C# usage throughout.

+++ Core/AeroVeloz.Domain/Entities/Users/Roles/Roles.cs
@@ -2,7 +2,12 @@
 
 namespace AeroVeloz.Domain.Entities.Users.Roles
 {
-    public partial class Roles : BEntity<int>
+    /// <summary>
+    /// Entidad de dominio que representa un rol dentro del sistema.
+    /// Los roles determinan el nivel de acceso y las acciones que un usuario puede realizar
+    /// (ej: SYSTEMADMIN, AIRPORTADMIN, AIRLINEADMIN).
+    /// </summary>
+    public partial class Roles : BEntity<short>
     {
         public string? nameRol { get; init; }
     }
